♻️ refactor(api): co-locate the JS minifier in clean
All three minifiers now live in one namespace: minify for HTML,
minify_css for a stylesheet, and minify_js for a script, each with its
frozen config object, sit together under turbohtml.clean. The
serializer layout configs Minify, Indent, and Formatter stay at the
package root, where the serialize surface uses them.

part of #478
