+++ docs/changelog/435.bugfix.rst
@@ -1,2 +1,2 @@
-:func:`~turbohtml.minify_js` no longer folds ``test?y:y`` to ``y`` when ``test`` is a bare identifier, which would drop
-the ``ReferenceError`` an unresolvable read throws.
+:func:`~turbohtml.clean.minify_js` no longer folds ``test?y:y`` to ``y`` when ``test`` is a bare identifier, which would
+drop the ``ReferenceError`` an unresolvable read throws.
