+++ docs/changelog/438.bugfix.rst
@@ -1,3 +1,3 @@
-:func:`~turbohtml.minify_js` now compares identifiers by their decoded ECMA-262 StringValue, so a Unicode-escaped
+:func:`~turbohtml.clean.minify_js` now compares identifiers by their decoded ECMA-262 StringValue, so a Unicode-escaped
 declaration links to its plainly-spelled use instead of being judged unused and dropped (turning the use into a free
 variable that throws).
