+++ docs/changelog/466.feature.rst
@@ -1,3 +1,3 @@
-Add an ``on_error`` keyword to :func:`~turbohtml.minify_js`. The default ``"raise"`` keeps failing on a script the
+Add an ``on_error`` keyword to :func:`~turbohtml.clean.minify_js`. The default ``"raise"`` keeps failing on a script the
 parser cannot handle; ``"passthrough"`` returns the source unchanged instead, the never-fail contract ``jsmin`` and
 ``rjsmin`` guarantee, reusing the verbatim fallback the inline-``<script>`` path already applies.
