+++ docs/how-to/serializing.rst
@@ -82,16 +82,17 @@ When the input is a string rather than a built tree -- the ``minify-html`` and `
  Minify JavaScript
 *******************
 
-:func:`~turbohtml.minify_js` minifies a JavaScript string on its own. It always folds whitespace, comments and number
-literals — ``/*! ... */`` bang comments and any ``@license`` / ``@preserve`` comment are the exception, kept byte-exact
-as a leading banner so a license header survives, exactly as the CSS minifier keeps them. A :class:`~turbohtml.JSMinify`
-toggles the two heavier passes — ``mangle`` renames local bindings to short names and ``fold`` runs constant folding and
-dead-code elimination. Top-level names are global, so they are never renamed; only bindings local to a function are. A
-construct the parser does not handle raises :class:`ValueError` rather than passing through silently.
+:func:`~turbohtml.clean.minify_js` minifies a JavaScript string on its own. It always folds whitespace, comments and
+number literals — ``/*! ... */`` bang comments and any ``@license`` / ``@preserve`` comment are the exception, kept
+byte-exact as a leading banner so a license header survives, exactly as the CSS minifier keeps them. A
+:class:`~turbohtml.clean.JSMinify` toggles the two heavier passes — ``mangle`` renames local bindings to short names and
+``fold`` runs constant folding and dead-code elimination. Top-level names are global, so they are never renamed; only
+bindings local to a function are. A construct the parser does not handle raises :class:`ValueError` rather than passing
+through silently.
 
 .. testcode::
 
-    from turbohtml import minify_js, JSMinify
+    from turbohtml.clean import minify_js, JSMinify
 
     source = "function f(x) { var half = x / 2; return half * half; }"
     print(minify_js(source))
@@ -102,7 +103,7 @@ construct the parser does not handle raises :class:`ValueError` rather than pass
     function f(b){var a=b/2;return a*a}
     function f(x){var half=x/2;return half*half}
 
-Inline ``<script>`` minification rides on HTML minification: pass a :class:`~turbohtml.JSMinify` as
+Inline ``<script>`` minification rides on HTML minification: pass a :class:`~turbohtml.clean.JSMinify` as
 :class:`~turbohtml.Minify`'s ``minify_js`` (the default ``None`` leaves scripts untouched). Only scripts the ``type``
 attribute marks as JavaScript are rewritten — a ``type="application/json"`` or ``importmap`` payload is left
 byte-for-byte — and a script the parser cannot handle is emitted verbatim, so one bad ``<script>`` never breaks the
@@ -111,7 +112,8 @@ document.
 .. testcode::
 
     import turbohtml
-    from turbohtml import Html, Minify, JSMinify
+    from turbohtml import Html, Minify
+    from turbohtml.clean import JSMinify
 
     doc = turbohtml.parse("<p>hi<script>function plus(a, b) { return a + b; }</script>")
     print(doc.serialize(Html(layout=Minify(minify_js=JSMinify()))))
