+++ docs/reference/serialize.rst
@@ -10,15 +10,13 @@ Turn a tree back into markup or text. Each renderer takes one configuration obje
 bounded ``str`` chunks for a large page (every layout but :class:`Minify`, which needs the whole tree);
 :meth:`Node.to_markdown` takes a :class:`Markdown` config; and :meth:`Node.to_text` and :meth:`Node.to_annotated_text`
 take a :class:`PlainText` config. :func:`escape` and :func:`unescape` are the standalone string helpers;
-:func:`annotation_surface` and :func:`annotation_tags` post-process the annotated-text result. :func:`minify_js`
-minifies a JavaScript string on its own, and a :class:`JSMinify` passed to :class:`Minify` extends HTML minification
-into inline ``<script>`` content.
+:func:`annotation_surface` and :func:`annotation_tags` post-process the annotated-text result. A
+:class:`~turbohtml.clean.JSMinify` passed to :class:`Minify` extends HTML minification into inline ``<script>`` content;
+the standalone :func:`~turbohtml.clean.minify_js` lives with the other minifiers in :mod:`turbohtml.clean`.
 
 .. autofunction:: escape
 
 .. autofunction:: unescape
-
-.. autofunction:: minify_js
 
 .. autoclass:: Html
     :members:
@@ -30,9 +28,6 @@ into inline ``<script>`` content.
     :members:
 
 .. autoclass:: Minify
-    :members:
-
-.. autoclass:: JSMinify
     :members:
 
 .. autoclass:: Markdown
