+++ tests/serialize/js/test_minify.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 
 import pytest
 
-from turbohtml import JSMinify, minify_js
+from turbohtml.clean import JSMinify, minify_js
 
 
 def minify(source: str) -> str:
